+++ wetmap/src/components/authentication/Landning/view.tsx
@@ -1,15 +1,15 @@
 import React, { Dispatch, SetStateAction } from 'react';
 // import style from './style.module.scss';
-import Button from '../../../reusables/button';
+import Button from '../../reusables/button';
 import blackManta from '../../../../images/blackManta.png';
 import googleIcon from '../../../../images/google-color.png';
 import facebookIcon from '../../../../images/facebook-color.png';
 import appleIcon from '../../../../images/apple.png';
-import ButtonIcon from '../../../reusables/buttonIcon';
+import ButtonIcon from '../../reusables/buttonIcon';
 import carouselData from '../carousel-data.json';
 import { LoginSocialApple, LoginSocialFacebook, LoginSocialGoogle } from 'reactjs-social-login';
-import { ActiveProfile } from '../../../../entities/profile';
-import WavyBlock from '../../../reusables/wavyBlock';
+import { ActiveProfile } from '../../../entities/profile';
+import WavyBlock from '../../reusables/wavyBlock';
 
 type LandingPageProps = {
   goToSlide:           (pageNumber: number) => void
